requirements.txt
flask 
flask-cors 
tensorflow
pydicom 
opencv-python 
pillow 
numpy
pylibjpeg 
pylibjpeg-libjpeg 
pylibjpeg-openjpeg